chore: rename arbitrum testnet to rinkarby
