+++ typescript/optics-deploy/scripts/alfajoresArbitrum/core.ts
@@ -1,6 +1,6 @@
 import { deployTwoChains } from '../../src/core';
 import * as alfajores from '../../config/testnets/alfajores';
-import * as arbitrum from '../../config/testnets/arbitrum';
+import * as rinkarby from '../../config/testnets/rinkarby';
 import { CoreDeploy } from '../../src/core/CoreDeploy';
 import { deployEnvironment } from '../../src/chain';
 
@@ -8,10 +8,10 @@ let environment = deployEnvironment();
 
 let alfaConfig =
   environment === 'staging' ? alfajores.stagingConfig : alfajores.devConfig;
-let arbitrumConfig =
-  environment === 'staging' ? arbitrum.stagingConfig : arbitrum.devConfig;
+let rinkarbyConfig =
+  environment === 'staging' ? rinkarby.stagingConfig : rinkarby.devConfig;
 
 const alfaDeploy = new CoreDeploy(alfajores.chain, alfaConfig);
-const arbitrumDeploy = new CoreDeploy(arbitrum.chain, arbitrumConfig);
+const rinkarbyDeploy = new CoreDeploy(rinkarby.chain, rinkarbyConfig);
 
-deployTwoChains(alfaDeploy, arbitrumDeploy);
+deployTwoChains(alfaDeploy, rinkarbyDeploy);
